+++ server/Src/Application/DTOs/Requests/SendCommandRequest.cs
@@ -4,6 +4,6 @@ public record SendCommandRequest
 {
     public required string Action { get; init; }  // "start" | "stop" | "setPitch" | "setInterval"
     public string? Reason { get; init; }           // stop only
-    public float? PitchAngle { get; init; }        // setPitch only
+    public double? PitchAngle { get; init; }        // setPitch only
     public int? IntervalSeconds { get; init; }     // setInterval only
 }
